FIX: add Correción de gráficas y lógica comparativa

+++ src/main/java/com/app/models/modeling/PuntoTemporal.java
@@ -4,5 +4,5 @@ package com.app.models.modeling;
  * Record para representar un punto en una serie temporal.
  * Utilizado para almacenar datos observados y modelados.
  */
-public record PuntoTemporal(int anio, double nacimientos) {
+public record PuntoTemporal(double anio, double nacimientos) {
 }
